remove _rezapi
